requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
loguru==0.7.2
